Tested out new raft level cross talk script.
